refactor: 테스트용 DB 세팅
