+++ src/main/java/janggi/Application.java
@@ -1,10 +1,11 @@
 package janggi;
 
-import janggi.dao.JanggiJdbcDao;
+import janggi.dao.GameJdbcDao;
+import janggi.dao.MoveJdbcDao;
 
 public class Application {
     public static void main(String[] args) {
-        final JanggiGame janggi = new JanggiGame(new JanggiJdbcDao());
+        final JanggiGame janggi = new JanggiGame(new GameJdbcDao(), new MoveJdbcDao());
         janggi.run();
     }
 }
