+++ src/main/java/janggi/dao/GameJdbcDao.java
@@ -2,39 +2,18 @@ package janggi.dao;
 
 import janggi.dto.MoveDto;
 import java.sql.Connection;
-import java.sql.DriverManager;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-public class JanggiJdbcDao implements JanggiDao {
-    private static final String SERVER = "localhost:13306"; // MySQL 서버 주소
-    private static final String DATABASE = "janggi"; // MySQL DATABASE 이름
-    private static final String OPTION = "?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC";
-    private static final String USERNAME = "root"; //  MySQL 서버 아이디
-    private static final String PASSWORD = "root"; // MySQL 서버 비밀번호
-
-    private final Connection connection;
-
-    public JanggiJdbcDao() {
-        this.connection = getConnection();
-    }
-
-    public Connection getConnection() {
-        try {
-            return DriverManager.getConnection("jdbc:mysql://" + SERVER + "/" + DATABASE + OPTION, USERNAME, PASSWORD);
-        } catch (final SQLException e) {
-            System.err.println("DB 연결 오류:" + e.getMessage());
-            e.printStackTrace();
-            return null;
-        }
-    }
+public class GameJdbcDao implements GameDao {
 
     @Override
     public void saveInitialGame(final int setupOption) {
         final String gameSaveQuery = "INSERT INTO game(setup_option, finished) value (?, false)";
-        try (final var preparedStatement = connection.prepareStatement(gameSaveQuery)) {
+        try (final Connection connection = ConnectionGenerator.getConnection();
+             final var preparedStatement = connection.prepareStatement(gameSaveQuery)) {
             preparedStatement.setInt(1, setupOption);
             preparedStatement.executeUpdate();
         } catch (final SQLException e) {
@@ -45,7 +24,8 @@ public class JanggiJdbcDao implements JanggiDao {
     @Override
     public boolean existNotFinishedGame() {
         final String notFinishedGameQuery = "SELECT * FROM game WHERE finished=0";
-        try (final var preparedStatement = connection.prepareStatement(notFinishedGameQuery)) {
+        try (final Connection connection = ConnectionGenerator.getConnection();
+             final var preparedStatement = connection.prepareStatement(notFinishedGameQuery)) {
             final ResultSet resultSet = preparedStatement.executeQuery();
             return resultSet.next();
         } catch (SQLException e) {
@@ -56,7 +36,8 @@ public class JanggiJdbcDao implements JanggiDao {
     @Override
     public int findRecentNotFinishedGameId() {
         final String notFinishedGameQuery = "SELECT * FROM game WHERE finished = 0 ORDER BY id DESC";
-        try (final var preparedStatement = connection.prepareStatement(notFinishedGameQuery)) {
+        try (final Connection connection = ConnectionGenerator.getConnection();
+             final var preparedStatement = connection.prepareStatement(notFinishedGameQuery)) {
             final ResultSet resultSet = preparedStatement.executeQuery();
             if (resultSet.next()) {
                 return resultSet.getInt("id");
@@ -71,7 +52,8 @@ public class JanggiJdbcDao implements JanggiDao {
     public List<Integer> findNotFinishedGameIds() {
         final String notFinishedGameQuery = "SELECT * FROM game WHERE finished=0";
         final List<Integer> ids = new ArrayList<>();
-        try (final var preparedStatement = connection.prepareStatement(notFinishedGameQuery)) {
+        try (final Connection connection = ConnectionGenerator.getConnection();
+             final var preparedStatement = connection.prepareStatement(notFinishedGameQuery)) {
             final ResultSet resultSet = preparedStatement.executeQuery();
             while (resultSet.next()) {
                 ids.add(resultSet.getInt("id"));
@@ -85,7 +67,8 @@ public class JanggiJdbcDao implements JanggiDao {
     @Override
     public int findGameSetup(final int gameId) {
         final String notFinishedGameQuery = "SELECT * FROM game WHERE id=?";
-        try (final var preparedStatement = connection.prepareStatement(notFinishedGameQuery)) {
+        try (final Connection connection = ConnectionGenerator.getConnection();
+             final var preparedStatement = connection.prepareStatement(notFinishedGameQuery)) {
             preparedStatement.setInt(1, gameId);
             final ResultSet resultSet = preparedStatement.executeQuery();
             if (resultSet.next()) {
@@ -97,41 +80,44 @@ public class JanggiJdbcDao implements JanggiDao {
         }
     }
 
-    @Override
-    public List<MoveDto> selectAllHistory(final int gameId) {
-        final String selectAllQuery = "SELECT start_row, start_column, end_row, end_column FROM moveHistory WHERE game_id = ?";
-        final List<MoveDto> moveDtos = new ArrayList<>();
-        try (final var preparedStatement = connection.prepareStatement(selectAllQuery)) {
-            preparedStatement.setInt(1, gameId);
-            final ResultSet historyResult = preparedStatement.executeQuery();
-            while (historyResult.next()) {
-                moveDtos.add(makeHistory(historyResult));
-            }
-        } catch (SQLException e) {
-            throw new RuntimeException("해당 게임의 움직인 기록들을 불러외지 못했습니다.");
-        }
-        return moveDtos;
-    }
-
-    @Override
-    public void saveHistory(final MoveDto moveDto, final int gameId) {
-        final String historySaveQuery = "INSERT INTO moveHistory(game_id, start_row, start_column, end_row, end_column) VALUES (?,?,?,?,?)";
-        try (final var preparedStatement = connection.prepareStatement(historySaveQuery)) {
-            preparedStatement.setInt(1, gameId);
-            preparedStatement.setString(2, moveDto.getStartRow());
-            preparedStatement.setString(3, moveDto.getStartColumn());
-            preparedStatement.setString(4, moveDto.getEndRow());
-            preparedStatement.setString(5, moveDto.getEndColumn());
-            preparedStatement.executeUpdate();
-        } catch (final SQLException e) {
-            throw new RuntimeException("움직인 기록을 저장하지 못했습니다.");
-        }
-    }
+//    @Override
+//    public List<MoveDto> selectAllHistory(final int gameId) {
+//        final String selectAllQuery = "SELECT start_row, start_column, end_row, end_column FROM moveHistory WHERE game_id = ?";
+//        final List<MoveDto> moveDtos = new ArrayList<>();
+//        try (final Connection connection = ConnectionGenerator.getConnection();
+//             final var preparedStatement = connection.prepareStatement(selectAllQuery)) {
+//            preparedStatement.setInt(1, gameId);
+//            final ResultSet historyResult = preparedStatement.executeQuery();
+//            while (historyResult.next()) {
+//                moveDtos.add(makeHistory(historyResult));
+//            }
+//        } catch (SQLException e) {
+//            throw new RuntimeException("해당 게임의 움직인 기록들을 불러외지 못했습니다.");
+//        }
+//        return moveDtos;
+//    }
+//
+//    @Override
+//    public void saveHistory(final MoveDto moveDto, final int gameId) {
+//        final String historySaveQuery = "INSERT INTO moveHistory(game_id, start_row, start_column, end_row, end_column) VALUES (?,?,?,?,?)";
+//        try (final Connection connection = ConnectionGenerator.getConnection();
+//             final var preparedStatement = connection.prepareStatement(historySaveQuery)) {
+//            preparedStatement.setInt(1, gameId);
+//            preparedStatement.setString(2, moveDto.getStartRow());
+//            preparedStatement.setString(3, moveDto.getStartColumn());
+//            preparedStatement.setString(4, moveDto.getEndRow());
+//            preparedStatement.setString(5, moveDto.getEndColumn());
+//            preparedStatement.executeUpdate();
+//        } catch (final SQLException e) {
+//            throw new RuntimeException("움직인 기록을 저장하지 못했습니다.");
+//        }
+//    }
 
     @Override
     public void setGameFinished(final int gameId) {
         final String gameSetFinishedQuery = "UPDATE game SET finished = true WHERE id = ?";
-        try (final var preparedStatement = connection.prepareStatement(gameSetFinishedQuery)) {
+        try (final Connection connection = ConnectionGenerator.getConnection();
+             final var preparedStatement = connection.prepareStatement(gameSetFinishedQuery)) {
             preparedStatement.setInt(1, gameId);
             preparedStatement.executeUpdate();
         } catch (final SQLException e) {
